Updated Czech translation
I've updated translation of file renaming messages and some items in paper sizes section. There is one thing - we aren't use US papers like "Letter", "Legat" etc. at all. So, we haven't any Czech words for these items. I can only translate "Half letter" to "Poloviční letter", which is bad translation - "Half letter" is the name of the paper size and it shouldn't be translated in a half way. So, it's not my bad there is missing some translations. They are actualy not missing :D.