Implement `--page-height` and `--page-width`
Issue #401

The original fix for #401 involved adding additional arguments for `scanimage` but globally. This had a bad side-effect for users with multiple scanners with differing characteristics.

The new solution treats these additional parameters as first class citizens and automatically adds them as required. The UI still does not present the user with the ability to override the values - although the API supports it should that become a requirement in the future. As a result the app always uses the default values provided by SANE.

The net result is that `scanimageAdditionalParameters` has been removed and is ignored. And no customisation should be required at all. However, in the event that the values provided by SANE are unsatisfactory, then it is possible to override the value as per all other device customisation - this has been noted in the `config.md` doc.

Other changes:
* Updated the yml used by swagger to refect the new parameters
* There is more granular checking of the individual geometry elements now - the previous behaviour relied on a bad assumption that `x`, `y`, `t` and `l` were always there or always not.
* Additional unit tests