Add description of pipeline mechanism
The pipeline commands aren't scripts. 
This PR makes it more clear in the docs.
E.g. using a shell variable in one line, it can't be accessed in another line. That is particular problematic, as empty shell variables silently return nothing. Commands might not execute or, even worse, completely wrong. E.g. rclone cp ./${foo} /path/desitination/ will copy the entire content of the tmp folder instead of the file which is named in foo.