Convert to classes
Converted most of the code to standard classes and removed dependencies on service instances (like `Config`). This makes each class much cleaner, and consistent and above all more unit testable on its own.

There is a new `Application` class which is a singleton service which provides factory methods for things like `config`, `context`, `deviceList` and all the stuff which happens at run time.

This impacted logging which should only be used after deliberate initialisation, but which was being prematurely initialised - so these have had their `require()`s moved inline with method calls where necessary.

This also tidies up a lot of unit tests, file names, class names and so on.