Various logging changes
No associated issue. But the logs were kind of out of hand and not always helpful. The structure of the server side `Request` class (with its embedded `Context`) made it particularly grevious.

It's all just been tidied up a bit and made a bit more succinct and useful - so even a default of INFO (which it now is) is sufficient for most debugging purposes.