Fix `scanimage -A` parsing with `[advanced]`
Issue #564

There are cases where `scanimage -A` outputs an additional square bracketed item on a parameter line such as:

```
    --contrast -127..127 [0] [advanced]
```

The old regex only assumed this happened once - which was good enough for 6 years or so. It seems that it can happen twice.

The regex has been updated to have a non-greedy match for the first one (`default`) and there is a subsequent optional non-capturing group for the second, which for want of a better name is referred here as `meta`.

It turns out that this pattern has been seen before with `[read-only]` but that was benign since it was treated as a default to ignore. For that reason, the `enabled` flag has to be altered too.

Associated unit test added.