Additional contrast filters (#617)
Added a filter to increase contrast.

On my printer (a Canon Pixma, fwiw), default scan settings lead to
"washed out" images, and increasing the contrast generates better scans.
In my case, the `-contrast` imagemagick setting was not that great (even
when doubled for more effect). That's why I came up with this
half-custom `-level 10%,90%` command. It now works great, at least for
B/W text documents.
For the sake of completeness, I have added both filters in this MR. Not
sure whether you want to provide both to all users, or if you find it
confusing, in which case I'll let you choose the one to provide to all
your users!

I have included translations to languages I know, but I'm not a native
German speaker.

---------

Co-authored-by: jerome <software@froiss.art>