Upgrade to Vue3 and Vite
This is a significant update. The old Vue2 app would not build with node18 or above which was gradually becoming important to fix. Further, now that IE has been out of support for over a year, it's reasonable to stop supporting it and other old browsers.

It would have been possible to upgrade just Vite or just Vue3 but going to all the effort to do half a job didn't seem worth it; as well as just the more mundane "keeping up to date"

Vite offers significant advantages in terms of development speed - UI builds now take seconds rather than minutes, but it also means that building the app requires node 14 (ideally later) or above. That said, the app will *run* using node 10+.

The changes are rather too many to enumerate but fall into the following categories:

* Update all (client) package dependencies
* Update `main.js` to use the latest `vuetify` and `vue-i18n`
* Upgrade component usage
* Remove all traces of vue-cli and webpack
* Small UI theming changes
* GitHub workflows (now require later node)

This PR is not intended to change **any** behaviour, packaging, or anything else.