Create bootstrap script
The intention of this script is to separate it from the actual installer.

As things currently stand, it is impossible to move the installer script or reorganise the packaging because it will break the one-line install.

Further, it isn't possible to one-line install a specific version.

This bootstrapper fixes all that. All it does is find the specified release, downloads it, extracts it and then runs `./installer.sh -i`

Also updated the install instructions and cleaned up the installer script