Convert release to deb package
The installation is now done with a debian binary package. The package is created manually for the time being since doing so with a source package seemed too big a step.

The new installation structure is:

```
/etc/scanservjs/ -> config directory
/srv/scanservjs/ -> data directory
/usr/lib/scanservjs/ -> code
```

* All the old release tarball code has been removed
* Docker now also uses the deb package which somewhat unifies the installation process (and testing)
* One line install stays in `bootstrap.sh` and it reacts differently to `deb` and `tar.gz` files.

Other changes along the way:

The npm install now uses `npm clean-install` which uses the package lock only and is much faster - this meant having to update the package lock

Also, although the core code will still run in node 10, the associated problems with npm and the fairly awful legacy packages which are complately broken (back in buster), along with package.json compatibility mean that I've just decided that this now requires nodejs >= 12 - given that even node12 is long out of LTS, I think this is reasonable. This means that in Debian 11 Bullseye, and above, you can just do a plain install and it will all just work. If you do the same on Buster, it will complain about broken dependencies. In that instance you install node16 manually using nodesource and then the install works. See `./makedeb.sh` for more on this.

The old `app-assets` - the installer and systemd unit have gone. Along with some ignore files.

The rest is just tidying up, removing unused dependencies and references.