File list must include all files (#676)
Closes #624

I have adopted the approach that a user may wish to rename the extension
of the file and should be allowed to do so. As such I just removed the
extension filter from the file list.

It does allow users to erroneously remove an extension, but they can
change it back without shelling anywhere.

If this is unpopular, I may add a UI restriction, but I think this is
sufficient to resolve the issue.