Fix various `v3` snagging issues
Closes #686
Closes #692
Closes #693
Closes #694

This is mostly documentation. But it also defines the least awful way to resolve `require(...)` paths where a plugin (i.e. `config.local.js`) is in a symlinked directory.

Node uses the absolute path of a module to determine how `require` works. When the module is in a symlinked directory then things go bad.